[3787] Extract last seen method
