+++ stream-chat-android-ui-components/src/main/kotlin/io/getstream/chat/android/ui/common/extensions/User.kt
@@ -17,25 +17,26 @@
 package io.getstream.chat.android.ui.common.extensions
 
 import android.content.Context
-import android.text.format.DateUtils
 import io.getstream.chat.android.client.models.User
 import io.getstream.chat.android.ui.R
-import io.getstream.chat.android.ui.common.extensions.internal.EMPTY
-import io.getstream.chat.android.uiutils.extensions.isInLastMinute
+import io.getstream.chat.android.uiutils.extensions.getLastSeenText
 
+/**
+ * Returns a string describing the elapsed time since the user was online (was watching the channel).
+ *
+ * Depending on the elapsed time, the string can have one of the following formats:
+ * - Online
+ * - Last seen just now
+ * - Last seen 13 hours ago
+ *
+ * @param context The context to load string resources.
+ * @return A string that represents the elapsed time since the user was online.
+ */
 public fun User.getLastSeenText(context: Context): String {
-    if (online) {
-        return context.getString(R.string.stream_ui_user_status_online)
-    }
-
-    return (lastActive ?: updatedAt ?: createdAt)?.let {
-        if (it.isInLastMinute()) {
-            context.getString(R.string.stream_ui_user_status_last_seen_just_now)
-        } else {
-            context.getString(
-                R.string.stream_ui_user_status_last_seen,
-                DateUtils.getRelativeTimeSpanString(it.time).toString()
-            )
-        }
-    } ?: String.EMPTY
+    return getLastSeenText(
+        context = context,
+        onlineResId = R.string.stream_ui_user_status_online,
+        justNowResId = R.string.stream_ui_user_status_last_seen_just_now,
+        lastSeenResId = R.string.stream_ui_user_status_last_seen,
+    )
 }
